my wallets added | flaw fixed
