+++ index.html
@@ -25,9 +25,10 @@
     </div>
     <div class="wallets">
       <h2>Мои кошельки</h2>
+      
       <div class="cards">
       </div>
-      <a href="">Смотреть все кошельки</a>
+      <a href="/pages/wallets/">Смотреть все кошельки</a>
     </div>
     <h2>Последние транзакции</h2>
     <table width = "1200 ">
@@ -41,8 +42,10 @@
 
     </table>
   </div>
-
-  <script src="./main.js"></script>
+  <script type="module" src="./main.js"></script>
+  <script type="module" src="/modules/header.js"></script>
+  <script type="module" src="/modules/cards.js"></script>
+  
 
 </body>
