+++ modules/police.js
@@ -1,0 +1,5 @@
+import { user } from "./user_data";
+
+if (!user) {
+    location.assign("/pages/login")
+}
