+++ modules/user_data.js
@@ -1,0 +1,2 @@
+
+export let user = JSON.parse(localStorage.getItem("user")) || null
