+++ pages/wallets/index.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="/style.css">
+
+    <title>Document</title>
+</head>
+<body>
+    <header>
+        <div class="header">
+    
+        </div>
+    </header>
+    <div class="container">
+        <div class="myWallets">
+            <h1>Мои кошельки</h1>
+            <a href="">alexadams@google.com</a>
+        </div>
+        <div class="cards">
+
+        </div>
+        <div class="blue_btn">
+            Добавить
+        </div>
+    </div>
+    <script type="module" src="/modules/cards.js"></script>
+    <script src="./index.js"></script>
+    <script type="module" src="/module/user_data.js"></script>
+    <script type="module" src="/modules/header.js"></script>
+</body>
+</html>
